Code refactoring select columns list files
